restructured project directories
