+++ lib/screens/note.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:my_notes/note_provide.dart';
+import 'package:my_notes/provider/note_provide.dart';
 import 'package:provider/provider.dart';
 
 class Note extends StatelessWidget {
