+++ lib/screens/notes_list.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:my_notes/note_provide.dart';
+import 'package:my_notes/provider/note_provide.dart';
 import 'package:provider/provider.dart';
-import './navigation_container.dart';
+import '../util/navigation_container.dart';
 import 'dart:core';
 
 class NotesList extends StatelessWidget {
