+++ pubspec.yaml
@@ -32,7 +32,6 @@ dependencies:
   cupertino_icons: ^1.0.2
   firebase_auth: ^4.4.2
   firebase_core: ^2.10.0
-  firebase_ui_auth: ^1.2.2
   flutter:
     sdk: flutter
   flutter_slidable: ^2.0.0
